Refactor and timing fix
Refactored code and rearranged the timing encoding. The encoding was changed to make things easier to pipeline in the future and to implement timing alignment (with x_pos and y_pos) using less gates. Also (in simulation -- need real world testing), fixed the timing issue causing the rightmost pixels to appear on the left side of the screen. The timing parameters were rewritten for this and a new parameter called H_RES and V_RES has been added to take what was H_ACTIVE_C and V_ACTIVE_L's place. I also renamed some nets, and the parameters above, to better fit their function. Another timing parameter was introduced called RAM_PREFETCH_CLKS which is used to start loading data from memory into the output registers. This is required so that the output matches with the x and y position vectors. Lastly, the state machine in the top verilog file has been removed and now the registers are pre-initialized - further simplifying the full VGA control logic.