Redid logic
Rewrote logic for the timing by starting to do a pipeline but later realizing it could just be done with a shift register. The way the new logic works is that the hsync, vsync, and active signals are propagated through registers, like in a pipeline, but in order to keep the longest path one clock cycle, the buffer address count, buffer, palette counter/RAMs were connected directly to each other - without the pipe between stages. The RAM is synchronous and needs one clock cycle to read. The point of the signals being propagated through the registers makes the timing work out nicely. The problem with this method is that if extra clock cycles were introduced in datapath, then it will be more work to adjust the pipeline to sync the control and RGB color.